Refactor command parsing into own module
